backend/requirements.txt
# Web Framework
Flask==3.0.3                    
Flask-CORS==4.0.0               

# Database 
Flask-SQLAlchemy==3.1.1         
Flask-Migrate==4.0.7            
psycopg2-binary==2.9.9          

# Authentication 
Flask-JWT-Extended==4.6.0       
bcrypt==4.1.3    

# Email 
sendgrid==6.11.0                

# Image Uploads 
cloudinary==1.40.0              

# Validation 
marshmallow==3.21.3             
Flask-Marshmallow==1.2.0        

# Utilities
python-dotenv==1.0.1            
Pillow==10.3.0                  

# API Documentation 
flask-swagger-ui==4.11.1        

# Testing 
pytest==8.2.0                   
pytest-cov==5.0.0               
factory-boy==3.3.0              
Faker==25.0.0                   